[DataSource][Arrow] Prune pushed filters that access partition columns (#1285)
